Multiple bugfixes with EG value retention and type conversion (especially when using filters), better PriorityScore TODO:  Check if this fixes the inability to set a reasonable filter gain without recalcing Pow(), maybe even Sqrt? -Algorithm.cs:  Intent is no longer hard-reset when operator size changes, only when a new Algorithm is made -Algorithm.cs:  Basic methods added to be used in the future for fixing connections when setting opCount and removing operators -Voice.cs:  Method to inform PriorityScore which ops eventually reach output -Channel.cs:  More accurately calculates PriorityScore when ops are bypassed or filtered -SlotIndicator.gd:  Be more democratic about resizing the grid with respect to intents.  (TODO:  Let Algorithm.cs determine the default layout when changing opSize)

+++ FMCore/Envelope.cs
@@ -208,7 +208,7 @@ namespace gdsFM
         {
             try
             {
-                this.SetVal(property, unchecked((int) val));
+                this.SetVal(property, unchecked(val));
                 // GD.Print(String.Format("Set op{0}.eg.{1} to {2}.", opTarget, property, val));
             } catch(NullReferenceException) {
                 #if GODOT
